requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
GitPython==3.1.40
pyyaml==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
